backend/weather_api/serializers.py
from rest_framework import serializers
from .models import UserLog, UserSettings

class LogSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserLog
        fields = '__all__'

class UserSettingsSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserSettings
        fields = '__all__'