requirements.txt
aiofiles            ==  24.1.0
aiogram             ==  2.0
aiohappyeyeballs    ==  2.4.3
aiohttp             ==  3.10.9
aiosignal           ==  1.3.1
annotated-types     ==  0.7.0
asgiref             ==  3.8.1
attrs               ==  24.2.0
babel               ==  2.16.0
certifi             ==  2024.8.30
charset-normalizer  ==  3.4.0
dj-database-url     ==  2.2.0
Django              ==  5.1.2
django-filter       ==  24.3
djangorestframework ==  3.15.2
environs            ==  11.0.0
frozenlist          ==  1.4.1
greenlet            ==  3.1.1
idna                ==  3.10
magic-filter        ==  1.0.12
marshmallow         ==  3.22.0
multidict           ==  6.1.0
packaging           ==  24.1
pip                 ==  23.2.1
propcache           ==  0.2.0
psycopg2-binary     ==  2.9.9
pydantic            ==  2.9.2
pydantic_core       ==  2.23.4
python-dotenv       ==  1.0.1
requests            ==  2.32.3
SQLAlchemy          ==  2.0.35
sqlparse            ==  0.5.1
typing_extensions   ==  4.12.2
urllib3             ==  2.2.3
yarl                ==  1.14.0==  